requirements.txt
# Web框架相关
Flask==2.3.3
Flask-CORS==4.0.0

# 数据库
pymongo==4.5.0

# HTTP请求
requests==2.31.0

# 爬虫相关
selenium>=4.0.0
beautifulsoup4>=4.9.0
webdriver-manager>=3.8.0

# 任务调度
schedule==1.2.0
python-crontab>=2.6.0

# Web服务器
gunicorn==20.1.0

# 音频处理相关
ffmpeg-python==0.2.0
pydub==0.25.1

# 语音识别相关
SpeechRecognition==3.10.4
openai-whisper==20231117
torch>=1.9.0
torchaudio>=0.9.0

# 翻译相关
googletrans==3.1.0a0
translators==5.8.9

# 文件处理
werkzeug==2.3.7